+++ ConsoleApplications/ConsoleApplications/src/_4_data_table/_4_4_filter_data/FilterData.cs
@@ -1,9 +1,9 @@
 using System;
 using System.Data;
 
-namespace ConsoleApplications._4_datatables._4_8_clear;
+namespace ConsoleApplications._4_data_table._4_4_filter_data;
 
-public class Clear {
+public class FilterData {
     public static void main() {
         DataTable employees = new DataTable();
 
@@ -71,42 +71,9 @@ public class Clear {
             DateTime.Now
         );
 
-        printEmployees(
-            "Before Clear",
-            ref employees
-        );
-
-        employees.Clear();
-
-        printEmployees(
-            "After Clear",
-            ref employees
-        );
-    }
-
-    private static void printEmployees(
-        string        title,
-        ref DataTable employees
-    ) {
-        Console.WriteLine(
-            new string(
-                '-',
-                30
-            ) + Environment.NewLine
-        );
-
-        Console.WriteLine(
-            title + Environment.NewLine
-        );
-
-        if (employees.Rows.Count > 0)
-            foreach (DataRow employee in employees.Rows)
-                printEmployee(
-                    employee
-                );
-        else
-            Console.WriteLine(
-                "Nothing to Show!" + Environment.NewLine
+        foreach (DataRow employee in employees.Rows)
+            printEmployee(
+                employee
             );
 
         Console.WriteLine(
@@ -114,6 +81,55 @@ public class Clear {
                 '-',
                 30
             ) + Environment.NewLine
+        );
+
+        const string FILTER = "Country = 'Syria'";
+        DataRow[] filteredEmployees = employees.Select(
+            FILTER
+        );
+
+        foreach (DataRow filteredEmployee in filteredEmployees)
+            printEmployee(
+                filteredEmployee
+            );
+
+        int count = filteredEmployees.Length;
+        double totalSalaries = Convert.ToDouble(
+            employees.Compute(
+                "SUM(Salary)",
+                FILTER
+            )
+        );
+        double averageSalaries = Convert.ToDouble(
+            employees.Compute(
+                "AVG(Salary)",
+                FILTER
+            )
+        );
+        double minimumSalaries = Convert.ToDouble(
+            employees.Compute(
+                "MIN(Salary)",
+                FILTER
+            )
+        );
+        double maximumSalaries = Convert.ToDouble(
+            employees.Compute(
+                "MAX(Salary)",
+                FILTER
+            )
+        );
+
+        Console.Write(
+            "Count: {0}"            + Environment.NewLine +
+            "Total Salaries: {1}"   + Environment.NewLine +
+            "Average Salaries: {2}" + Environment.NewLine +
+            "Minimum Salaries: {3}" + Environment.NewLine +
+            "Maximum Salaries: {4}" + Environment.NewLine,
+            count,
+            totalSalaries,
+            averageSalaries,
+            minimumSalaries,
+            maximumSalaries
         );
     }
