+++ ConsoleApplications/ConsoleApplications/src/_4_data_table/_4_5_sorting/Sorting.cs
@@ -1,9 +1,9 @@
 using System;
 using System.Data;
 
-namespace ConsoleApplications._4_datatables._4_3_count_and_sum_and_avg_and_min_and_max;
+namespace ConsoleApplications._4_data_table._4_5_sorting;
 
-public class CountAndSumAndAvgAndMinAndMax {
+public class Sorting {
     public static void main() {
         DataTable employees = new DataTable();
 
@@ -51,17 +51,61 @@ public class CountAndSumAndAvgAndMinAndMax {
         );
         employees.Rows.Add(
             4,
+            "Mohamed",
+            "Syria",
+            5120,
+            DateTime.Now
+        );
+        employees.Rows.Add(
+            5,
             "Ibrahim",
             "Jordan",
             5100,
             DateTime.Now
         );
         employees.Rows.Add(
-            5,
+            6,
             "Mustafa",
             "Iraq",
             4950,
             DateTime.Now
+        );
+
+        printEmployees(
+            "Before Sort",
+            ref employees
+        );
+
+        employees.DefaultView.Sort = "Name";
+        employees                  = employees.DefaultView.ToTable();
+
+        printEmployees(
+            "After Sort by Name Ascending",
+            ref employees
+        );
+
+        employees.DefaultView.Sort = "ID DESC";
+        employees                  = employees.DefaultView.ToTable();
+
+        printEmployees(
+            "After Sort by ID Descending",
+            ref employees
+        );
+    }
+
+    private static void printEmployees(
+        string        title,
+        ref DataTable employees
+    ) {
+        Console.WriteLine(
+            new string(
+                '-',
+                30
+            ) + Environment.NewLine
+        );
+
+        Console.WriteLine(
+            title + Environment.NewLine
         );
 
         foreach (DataRow employee in employees.Rows)
@@ -73,46 +117,7 @@ public class CountAndSumAndAvgAndMinAndMax {
             new string(
                 '-',
                 30
-            )
-        );
-
-        int count = employees.Rows.Count;
-        double totalSalaries = Convert.ToDouble(
-            employees.Compute(
-                "SUM(Salary)",
-                string.Empty
-            )
-        );
-        double averageSalaries = Convert.ToDouble(
-            employees.Compute(
-                "AVG(Salary)",
-                string.Empty
-            )
-        );
-        double minimumSalaries = Convert.ToDouble(
-            employees.Compute(
-                "MIN(Salary)",
-                string.Empty
-            )
-        );
-        double maximumSalaries = Convert.ToDouble(
-            employees.Compute(
-                "MAX(Salary)",
-                string.Empty
-            )
-        );
-
-        Console.Write(
-            "Count: {0}"            + Environment.NewLine +
-            "Total Salaries: {1}"   + Environment.NewLine +
-            "Average Salaries: {2}" + Environment.NewLine +
-            "Minimum Salaries: {3}" + Environment.NewLine +
-            "Maximum Salaries: {4}" + Environment.NewLine,
-            count,
-            totalSalaries,
-            averageSalaries,
-            minimumSalaries,
-            maximumSalaries
+            ) + Environment.NewLine
         );
     }
